+++ packages/falcon-ecommerce-uikit/src/SignIn/SignInMutation.tsx
@@ -2,7 +2,7 @@ import gql from 'graphql-tag';
 import { Mutation } from 'react-apollo';
 
 export const SIGN_IN_MUTATION = gql`
-  mutation signIn($input: SignIn!) {
+  mutation SignIn($input: SignIn!) {
     signIn(input: $input)
   }
 `;
@@ -11,6 +11,6 @@ export class SignInMutation extends Mutation {
   static defaultProps = {
     mutation: SIGN_IN_MUTATION,
     awaitRefetchQueries: true,
-    refetchQueries: ['customer', 'miniAccount']
+    refetchQueries: ['Customer', 'MiniAccount', 'Cart']
   };
 }