+++ packages/falcon-ecommerce-uikit/src/SignOut/SignOutMutation.tsx
@@ -2,7 +2,7 @@ import gql from 'graphql-tag';
 import { Mutation } from 'react-apollo';
 
 export const SIGN_OUT_MUTATION = gql`
-  mutation signOut {
+  mutation SignOut {
     signOut
   }
 `;
@@ -11,6 +11,6 @@ export class SignOutMutation extends Mutation {
   static defaultProps = {
     mutation: SIGN_OUT_MUTATION,
     awaitRefetchQueries: true,
-    refetchQueries: ['customer', 'miniAccount']
+    refetchQueries: ['Customer', 'MiniAccount', 'Cart']
   };
 }